# android/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Phos API",
    "description": "AI-powered photo/video manager",
    "license": {
      "name": "MIT"
    },
    "version": "dev"
  },
  "paths": {
    "/api/auth/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Handle OIDC callback",
        "description": "Handles the OIDC provider callback after user authentication. Validates the CSRF token, exchanges the authorization code for tokens, verifies the ID token, and issues a session JWT cookie.",
        "operationId": "callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code from OIDC provider",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "CSRF state token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "description": "Error code from OIDC provider",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error_description",
            "in": "query",
            "description": "Error description from OIDC provider",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to app after successful auth"
          },
          "400": {
            "description": "Invalid callback parameters"
          },
          "502": {
            "description": "Token exchange or ID token verification failed"
          }
        }
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh session JWT",
        "description": "Accepts a currently-valid session (Authorization: Bearer header or session cookie) and issues a new session JWT with a fresh expiry. Expired or invalid sessions are rejected with 401 — this endpoint cannot resurrect an expired session. Cookie-authenticated callers also get the session cookie re-set.",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "New session JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Session missing, invalid, or expired"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange OIDC ID token for session JWT",
        "description": "Mobile clients perform the OIDC Authorization Code + PKCE flow directly with the identity provider, then send the resulting ID token here. Phos validates it against the provider's JWKS and returns a session JWT that can be used in the `Authorization: Bearer` header.",
        "operationId": "token_exchange",
        "requestBody": {
          "description": "OIDC ID token to exchange",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID token format"
          },
          "401": {
            "description": "ID token verification failed"
          }
        }
      }
    },
...
    "/api/faces/dedupe": {
      "post": {
        "tags": [
          "faces"
        ],
        "summary": "Remove duplicate face boxes",
        "description": "Collapse overlapping face detections of the same face across the library, so reviewers do not have to delete them by hand. Two boxes assigned to different people are never merged, and shots already marked reviewed are left alone. Pass `dry_run=true` to get the count without deleting anything.",
        "operationId": "dedupe_faces",
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "description": "Report what would go without deleting anything. Defaults to `false`.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sweep result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DedupeFacesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/faces/{id}": {
      "delete": {
        "tags": [
          "faces"
        ],
        "summary": "Delete a face",
        "description": "Delete a face detection record. Cleans up orphaned people and recalculates shot assignments.",
        "operationId": "delete_face",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Face ID to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Face deleted successfully"
          },
          "404": {
            "description": "Face not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/faces/{id}/person": {
      "put": {
        "tags": [
          "faces"
        ],
        "summary": "Reassign a face",
        "description": "Reassign a face to a different person. Updates the shot's primary person if needed.",
        "operationId": "reassign_face",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Face ID to reassign",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignFacePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Face reassigned successfully"
          },
          "404": {
            "description": "Face or target person not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/faces/{id}/suggestions": {
      "get": {
        "tags": [
          "faces"
        ],
        "summary": "Get face suggestions",
        "description": "Get person assignment suggestions for a face based on embedding similarity to existing people.",
        "operationId": "get_face_suggestions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Face ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of person suggestions sorted by similarity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FaceSuggestion"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Face not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/faces/{id}/thumbnail": {
      "get": {
        "tags": [
          "faces"
        ],
        "summary": "Get face thumbnail",
        "description": "Retrieve the cropped JPEG thumbnail image for a detected face.",
        "operationId": "get_face_thumbnail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Face ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JPEG face thumbnail",
            "content": {
              "image/jpeg": {}
            }
          },
          "404": {
            "description": "Face not found or source image missing"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/files/{id}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Download a file",
        "description": "Download the original file content by file ID. Returns the raw bytes with the appropriate content type.",
        "operationId": "get_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File content",
            "content": {
              "application/octet-stream": {}
            }
          },
          "404": {
            "description": "File not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "files"
        ],
        "summary": "Delete a file",
        "description": "Delete a non-original file from a shot. The original file cannot be deleted directly — delete the shot instead.",
        "operationId": "delete_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Cannot delete original file"
          },
          "404": {
            "description": "File not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/files/{id}/faces": {
      "post": {
        "tags": [
          "faces"
        ],
        "summary": "Add manual face",
        "description": "Manually add a face region to a file with a specified bounding box and person assignment.",
        "operationId": "add_manual_face",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID to add a manual face to",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddManualFacePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Face added successfully"
          },
          "400": {
            "description": "Invalid bounding box coordinates"
          },
          "404": {
            "description": "File not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/files/{id}/set-original": {
      "put": {
        "tags": [
          "files"
        ],
        "summary": "Set file as original",
        "description": "Mark a file as the original (primary) file for its shot, demoting the current original.",
        "operationId": "set_file_original",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "File not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/files/{id}/thumbnail": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Get file thumbnail",
        "description": "Retrieve a resized thumbnail or video preview for a file. Supports configurable dimensions via query parameters.",
        "operationId": "get_file_thumbnail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "w",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thumbnail JPEG image",
            "content": {
              "image/jpeg": {}
            }
          },
          "404": {
            "description": "File not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/import/finalize": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Finalize import",
        "description": "Finalize the import process by running a full library scan and reorganization on newly imported files.",
        "operationId": "finalize_import",
        "responses": {
          "200": {
            "description": "Import finalized successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/import/upload": {
      "put": {
        "tags": [
          "import"
        ],
        "summary": "Upload a file",
        "description": "Upload a raw file for import. The file is written to the import staging area and queued for analysis, which runs in the background. Poll `/api/import/status` to see when it is done. Supports up to 1 GB files.",
        "operationId": "upload_file_raw",
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Raw file bytes",
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "File stored and queued for analysis"
          },
          "400": {
            "description": "Invalid filename"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/organize/stats": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get organization statistics",
        "description": "Return detailed organization statistics including pending reviews, confirmed shots, unsorted shots, and unnamed people.",
        "operationId": "get_organize_stats",
        "responses": {
          "200": {
            "description": "Detailed organize statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizeStatsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/reorganize": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Trigger reorganization",
        "description": "Trigger a background filesystem reorganization that moves files into person-based folder structure.",
        "operationId": "trigger_reorganize",
        "responses": {
          "200": {
            "description": "Reorganization started",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DedupeFacesResponse": {
        "type": "object",
        "description": "Result of a duplicate-box sweep.",
        "required": [
          "removed",
          "dry_run"
        ],
        "properties": {
          "dry_run": {
            "type": "boolean",
            "description": "Whether anything was actually deleted."
          },
          "removed": {
            "type": "integer",
            "description": "Faces removed — or, for a dry run, faces that would be removed.",
            "minimum": 0
          }
        }
      },
...
      "TokenExchangeRequest": {
        "type": "object",
        "required": [
          "id_token"
        ],
        "properties": {
          "id_token": {
            "type": "string"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "description": "Response containing a Phos session JWT.",
        "required": [
          "token",
          "expires_in"
        ],
        "properties": {
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session JWT set as HttpOnly cookie (phos_session) after OIDC login"
      }
    }
  },
  "security": [
    {
      "session_cookie": []
    }
  ]
}